requirements.txt
flask
gradio
pytesseract
pillow 
google-generativeai